[feat] int-> long 변경(할인값)

+++ src/main/java/com/fouribnb/coupon/presentation/dto/response/CreateCouponResponseDto.java
@@ -14,7 +14,7 @@ public class CreateCouponResponseDto {
     private UUID paymentId;
     private String couponName;
     private CouponStatus couponStatus;
-    private int discountValue;
+    private Long discountValue;
     private Boolean isUsed;
 
 }
